payments/src/events/listeners/__tests__/order-cancelled-listener.test.ts
import { OrderCancelledEvent, OrderStatus } from "@farmmicro/common";
import { natsWrapper } from "../../../nats-wrapper";
import mongoose from "mongoose";
import { Message } from "node-nats-streaming";
import { Order } from "../../../model/orderModel";
import { OrderCancelledListener } from "../order-cancelled-listener";

const setup = async () => {
  const listener = new OrderCancelledListener(natsWrapper.client);

  const order = Order.build({
    id: new mongoose.Types.ObjectId().toHexString(),
    user: new mongoose.Types.ObjectId().toHexString(),
    totalPrice: 1080,
    status: OrderStatus.Created,
    version: 0,
  });

  await order.save();

  const data: OrderCancelledEvent["data"] = {
    id: order.id,
    version: 1,
    orderItems: [
      {
        name: "test name",
        qty: 10,
        price: 100,
        image: "image",
        productId: new mongoose.Types.ObjectId().toHexString(),
      },
    ],
  };

  //@ts-ignore

  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, data, msg };
};

it("Replicates the order info", async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  const updatedOrder = await Order.findById(data.id);
  expect(updatedOrder!.id).toEqual(data.id);
  expect(updatedOrder!.status).toEqual(OrderStatus.Cancelled);
});

it("acks the message", async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);
  expect(msg.ack).toHaveBeenCalled();
});
